requirements.txt
faiss-cpu
langchain
numpy
nltk
sentence_transformers
streamlit
langchain_googel_genai
json